Some changes to ActiveGame slice state
Removed the moveResutlQueue, and the pendingMoveResult message (the acknowledgeMoveResult action also becomes redundant). Instead, received MoveResultMessages are immediately processed and applied to the state.

Also added the sunkenShipSurroundings array to the GridState interface - supposed to hold coordinates which are considered to be empty since they are right next to a ship.

utils.ts - made the neccesary changes to accomodate the aforementioned state type changes