Change structure of redux store
Merged the GameRoom and ActiveGame slices into the single GameRoom slice. The state of the GameRoom slice can be one of two types: Inactive or Active. The Inactive version has most of its fields optional, while the Active version has all of the same fields as mandatory, and also keeps track of the players and opponents score.

Moved all of the async redux thunks into separate modules.

wsMiddleware: changed the types of actions handled by the middleware (now only listening for 'openWSConnection' and 'sendMessage' actions, and dispatches only 'messageReceived').