Change app routes structure
Created the root route '/game' with two subroutes: 'getReady' and 'play' (which were previously root routes).

Created the GameScreen component which renders for the '/game' route and renders an react-router Outlet for the sub-routes.

Moved the PlacementScreen and ActiveGameScreen components into the GameScreen component directory.

Also moved the PlacementGridContext module into the useShipDrag hook directory.