GameRoomSlice - fix initial state assignment
When creating the slice, assign a copy of the defined initial state to the slice initial state.

Later when reseting the state to its initial value, use Object.assign() to assign values from initial state directly to the existing slice state object - this way the Immer correctly tracks changes to the state.